refactor: Activty, Screen 파일 분리
